updated admin booking controller -> Approve/Reject

+++ src/main/java/com/vof/dto/request/PaymentProofRequest.java
@@ -5,10 +5,14 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
+import java.math.BigDecimal;
+
 @Data
 public class PaymentProofRequest {
     @NotBlank(message = "Booking ID is required") @Size(max = 64) private String bookingId;
     @NotBlank(message = "UTR number is required") @Size(min = 6, max = 64) private String utrNumber;
-    @NotNull(message = "Amount is required") @DecimalMin(value = "0.01") private Double amount;
+    @NotNull(message = "Amount is required")
+    @DecimalMin(value = "0.01")
+    private BigDecimal amount;
     private MultipartFile screenshot;
 }
